+++ tests/test_stl.cpp
@@ -53,6 +53,10 @@ struct NonAssignable {
   NonAssignable &operator=(const NonAssignable &) = delete;
 };
 
+struct NonDefaultConstructible : Movable {
+    NonDefaultConstructible(int v) : Movable(v) {}
+};
+
 struct StructWithReadonlyMap {
     std::map<std::string, uint64_t> map;
 };
@@ -117,6 +121,10 @@ NB_MODULE(test_stl_ext, m) {
     nb::class_<NonAssignable>(m, "NonAssignable")
         .def(nb::init<>())
         .def_rw("value", &NonAssignable::value);
+
+    nb::class_<NonDefaultConstructible>(m, "NonDefaultConstructible")
+        .def(nb::init<int>())
+        .def_rw("value", &NonDefaultConstructible::value);
 
     nb::class_<StructWithReadonlyMap>(m, "StructWithReadonlyMap")
         .def(nb::init<>())
@@ -281,6 +289,10 @@ NB_MODULE(test_stl_ext, m) {
     m.def("variant_ret_var_none", []() { return std::variant<std::monostate, Copyable, int>(); });
     m.def("variant_unbound_type", [](std::variant<std::monostate, nb::list, nb::tuple, int> &x) { return x; },
           nb::arg("x") = nb::none());
+    m.def("variant_nondefault",
+          [](std::variant<NonDefaultConstructible, int> v) {
+              return v.index() == 0 ? std::get<0>(v).value : -std::get<1>(v);
+          });
 
     // ----- test50-test57 ------
     m.def("map_return_movable_value", [](){